+++ src/steps/resource-manager/security/index.ts
@@ -14,9 +14,15 @@ import {
   SecurityRelationships,
   SecuritySteps,
 } from './constants';
-import { createAssessmentEntity } from './converters';
+import {
+  createAssessmentEntity,
+  createSecurityContactEntity,
+} from './converters';
 import { STEP_RM_RESOURCES_RESOURCE_GROUPS } from '../resources';
-import { SUBSCRIPTION_ENTITY_METADATA } from '../subscriptions';
+import {
+  SUBSCRIPTION_ENTITY_METADATA,
+  STEP_RM_SUBSCRIPTIONS,
+} from '../subscriptions/constants';
 export * from './constants';
 
 export async function fetchAssessments(
@@ -56,6 +62,36 @@ export async function fetchAssessments(
   );
 }
 
+export async function fetchSecurityCenterContacts(
+  executionContext: IntegrationStepContext,
+): Promise<void> {
+  const { instance, logger, jobState } = executionContext;
+  const accountEntity = await jobState.getData<Entity>(ACCOUNT_ENTITY_TYPE);
+
+  const webLinker = createAzureWebLinker(accountEntity.defaultDomain as string);
+  const client = new SecurityClient(instance.config, logger);
+
+  const subscriptionId = `/subscriptions/${instance.config.subscriptionId}`;
+  const subscriptionEntity = await jobState.findEntity(subscriptionId);
+
+  await client.iterateSecurityContacts(async (securityContact) => {
+    const securityContactEntity = await jobState.addEntity(
+      createSecurityContactEntity(webLinker, securityContact),
+    );
+
+    if (!subscriptionEntity) return;
+
+    await jobState.addRelationship(
+      createDirectRelationship({
+        _class:
+          SecurityRelationships.SUBSCRIPTION_HAS_SECURITY_CENTER_CONTACT._class,
+        from: subscriptionEntity,
+        to: securityContactEntity,
+      }),
+    );
+  });
+}
+
 export const securitySteps: Step<
   IntegrationStepExecutionContext<IntegrationConfig>
 >[] = [
@@ -66,5 +102,15 @@ export const securitySteps: Step<
     relationships: [SecurityRelationships.SUBSCRIPTION_PERFORMED_ASSESSMENT],
     dependsOn: [STEP_AD_ACCOUNT, STEP_RM_RESOURCES_RESOURCE_GROUPS],
     executionHandler: fetchAssessments,
+  },
+  {
+    id: SecuritySteps.SECURITY_CENTER_CONTACTS,
+    name: 'Security Contacts',
+    entities: [SecurityEntities.SECURITY_CENTER_CONTACT],
+    relationships: [
+      SecurityRelationships.SUBSCRIPTION_HAS_SECURITY_CENTER_CONTACT,
+    ],
+    dependsOn: [STEP_AD_ACCOUNT, STEP_RM_SUBSCRIPTIONS],
+    executionHandler: fetchSecurityCenterContacts,
   },
 ];